Python/requirements.txt
python-osc
python-escpos
pyusb
Pillow